MCU_43 - Separated the internal-use-only members of AuxiliaryUpgradeConsole into a separate partial class file
